Refactor: Changing managerInterface location
